Move methods for syncing to syncer
